fix: renamed Hooks/ folder to SinchEvents/
